FIX: Pip 20+ support and fix tests
Apparently, the tests did not run since the
testenv did not match, so the builds were green
no matter the code.

I changed the tox.ini and added py38-pip20 and
made sure it failed before fixing the imports
in pkgversion.py.

On another note, we should probably not use
internals like this because it will randomly
break (like now). Fixing it now so we are
unblocked at least.